testing lookup

+++ labs/fibvec/FibVec.h
@@ -28,7 +28,7 @@ public:
   size_t capacity() const; // return size of storage buffer DONE
   size_t count() const; // return num of items in vector DONE
   void insert(int val, size_t index); // inserts val at index DONE
-  int lookup(size_t index) const; // returns val at index DONE
+  int lookup(int index) const; // returns val at index DONE
   int pop(); // removes and returns last value of vector
   void push(int val); // inserts val at end of vector DONE
   int remove(size_t index);
